requirements.txt
fastapi[standard]
uvicorn
pydantic
httpx
python-dotenv
pandas
numpy
motor
pymongo
transformers
torch
torchvision
python-jose
python-multipart
langchain
langchain-core
langgraph
pyjwt